move validate label to label store
