+++ bentoml/yatai/client/bento_repository_api.py
@@ -17,14 +17,14 @@
 import io
 import os
 import logging
-import re
 import tarfile
 import requests
 import shutil
 
 
-from bentoml.exceptions import BentoMLException, InvalidArgument
+from bentoml.exceptions import BentoMLException
 from bentoml.yatai.client.label_utils import generate_gprc_labels_selector
+from bentoml.yatai.label_store import _validate_labels
 from bentoml.yatai.proto.repository_pb2 import (
     AddBentoRequest,
     GetBentoRequest,
@@ -41,38 +41,6 @@ from bentoml.yatai.status import Status
 
 
 logger = logging.getLogger(__name__)
-
-
-def _validate_labels(labels):
-    """
-    Validate labels key value format is:
-        * Between 3 and 63 characters
-        * Consist of alphanumeric, dash (-), period (.), and underscore (_)
-        * Start and end with alphanumeric
-    Args:
-        labels: Dictionary
-
-    Returns:
-    Raise:
-        InvalidArgument
-    """
-    if not isinstance(labels, dict):
-        raise InvalidArgument('BentoService labels must be a dictionary')
-
-    pattern = re.compile("^(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?$")
-    for key in labels:
-        if (
-            not (63 >= len(key) >= 3)
-            or not (63 >= len(labels[key]) >= 3)
-            or not pattern.match(key)
-            or not pattern.match(labels[key])
-        ):
-            raise InvalidArgument(
-                f'Invalide label {key}:{labels[key]}. Valid label key and value must '
-                f'be between 3 to 63 characters and must be begin and end with '
-                f'an alphanumeric character ([a-z0-9A-Z]) with dashes (-), '
-                f'underscores (_), and dots (.).'
-            )
 
 
 class BentoRepositoryAPIClient:
