+++ bentoml/yatai/label_store.py
@@ -1,7 +1,9 @@
 import enum
+import re
+
 from sqlalchemy import UniqueConstraint, Column, Integer, String, and_, Enum
 
-from bentoml.exceptions import YataiLabelException
+from bentoml.exceptions import YataiLabelException, InvalidArgument
 from bentoml.yatai.db import Base
 from bentoml.yatai.proto.label_selectors_pb2 import LabelSelectors
 
@@ -141,3 +143,35 @@ def filter_label_query(sess, resource_type, label_selectors):
     query = query.intersect(*filters)
     result = query.all()
     return [row[0] for row in result]
+
+
+def _validate_labels(labels):
+    """
+    Validate labels key value format is:
+        * Between 3 and 63 characters
+        * Consist of alphanumeric, dash (-), period (.), and underscore (_)
+        * Start and end with alphanumeric
+    Args:
+        labels: Dictionary
+
+    Returns:
+    Raise:
+        InvalidArgument
+    """
+    if not isinstance(labels, dict):
+        raise InvalidArgument('BentoService labels must be a dictionary')
+
+    pattern = re.compile("^(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?$")
+    for key in labels:
+        if (
+            not (63 >= len(key) >= 3)
+            or not (63 >= len(labels[key]) >= 3)
+            or not pattern.match(key)
+            or not pattern.match(labels[key])
+        ):
+            raise InvalidArgument(
+                f'Invalide label {key}:{labels[key]}. Valid label key and value must '
+                f'be between 3 to 63 characters and must be begin and end with '
+                f'an alphanumeric character ([a-z0-9A-Z]) with dashes (-), '
+                f'underscores (_), and dots (.).'
+            )
